move back get_otel_data_exporter
