+++ shared/observability/src/airflow_shared/observability/metrics/otel_logger.py
@@ -32,6 +32,7 @@ from opentelemetry.sdk.metrics._internal.export import (
 )
 from opentelemetry.sdk.resources import SERVICE_NAME, Resource
 
+from ..common import get_otel_data_exporter
 from ..otel_env_config import load_metrics_env_config
 from .protocols import Timer
 from .validators import (
@@ -44,8 +45,6 @@ from .validators import (
 
 if TYPE_CHECKING:
     from opentelemetry.metrics import Instrument
-    from opentelemetry.sdk.metrics._internal.export import MetricExporter
-    from opentelemetry.sdk.trace.export import SpanExporter
     from opentelemetry.util.types import Attributes
 
     from .protocols import DeltaType
@@ -385,81 +384,6 @@ def flush_otel_metrics():
 
 def atexit_register_metrics_flush():
     atexit.register(flush_otel_metrics)
-
-
-def get_otel_data_exporter(
-    *,
-    otel_env_config: OtelEnvConfig,
-    host: str | None = None,
-    port: int | None = None,
-    ssl_active: bool = False,
-) -> SpanExporter | MetricExporter:
-    protocol = "https" if ssl_active else "http"
-
-    # According to the OpenTelemetry Spec, specific config options like 'OTEL_EXPORTER_OTLP_TRACES_ENDPOINT'
-    # take precedence over generic ones like 'OTEL_EXPORTER_OTLP_ENDPOINT'.
-    env_exporter_protocol = (
-        otel_env_config.type_specific_exporter_protocol or otel_env_config.exporter_protocol
-    )
-    env_endpoint = otel_env_config.type_specific_endpoint or otel_env_config.base_endpoint
-
-    # If the protocol env var isn't set, then it will be None,
-    # and it will default to an http/protobuf exporter.
-    if env_endpoint and env_exporter_protocol == "grpc":
-        from opentelemetry.exporter.otlp.proto.grpc.metric_exporter import OTLPMetricExporter
-        from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
-    else:
-        from opentelemetry.exporter.otlp.proto.http.metric_exporter import OTLPMetricExporter
-        from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
-
-    if env_endpoint:
-        if host is not None and port is not None:
-            log.warning(
-                "Both the standard OpenTelemetry environment variables and "
-                "the Airflow OpenTelemetry configs have been provided. "
-                "Using the OpenTelemetry environment variables. "
-                "The Airflow configs have been deprecated and will be removed in the future."
-            )
-
-        endpoint_str = env_endpoint
-        # The SDK will pick up all the values from the environment.
-        if otel_env_config.data_type == OtelDataType.TRACES:
-            exporter = OTLPSpanExporter()
-        else:
-            exporter = OTLPMetricExporter()
-    else:
-        if host is None or port is None:
-            # Since the configs have been deprecated, host and port could be None.
-            # Log a warning to steer the user towards configuring the environment variables
-            # and deliberately let it fail here without providing fallbacks.
-            log.warning(
-                "OpenTelemetry %s have been enabled but the endpoint settings haven't been configured. "
-                "The Airflow configs have been deprecated and will be removed in the future. "
-                "Configure the standard OpenTelemetry environment variables instead. "
-                "For more info, check the docs.",
-                otel_env_config.data_type.value,
-            )
-        else:
-            log.warning(
-                "The Airflow OpenTelemetry configs have been deprecated and will be removed in the future. "
-                "Configure the standard OpenTelemetry environment variables instead. "
-                "For more info, check the docs."
-            )
-        # If the environment endpoint isn't set, then assume that the airflow config is used
-        # where protocol isn't specified, and it's always http/protobuf.
-        # In that case it should default to the full 'url_path' and set it directly.
-
-        endpoint_suffix = "traces" if otel_env_config.data_type == OtelDataType.TRACES else "metrics"
-
-        endpoint_str = f"{protocol}://{host}:{port}/v1/{endpoint_suffix}"
-        if otel_env_config.data_type == OtelDataType.TRACES:
-            exporter = OTLPSpanExporter(endpoint=endpoint_str)
-        else:
-            exporter = OTLPMetricExporter(endpoint=endpoint_str)
-
-    exporter_name = exporter.__class__.__name__
-    log.info("Connecting to OpenTelemetry Collector", exporter_name=exporter_name, endpoint=endpoint_str)
-    return exporter
 
 
 def get_otel_logger(
